requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.21
beautifulsoup4==4.14.3
google-genai==1.57.0
weasyprint==67.0
fpdf2==2.8.5
Jinja2==3.1.6
pandas==2.3.3
pyarrow==22.0.0
python-dotenv==1.2.1
websockets==15.0.1
PyMuPDF==1.26.7
supabase==2.27.0
selenium==4.41.0
winrt-runtime==3.2.1
winsdk==1.0.0b10
google-auth==2.47.0
google-auth-oauthlib==1.2.4
google-api-python-client==2.190.0